Refactor type checking of user-set variable
Previously, the function which checked that g:decfile_disable_folding
was of type number both checked the type of the variable, and printed an
error message if the type was incorrect. In retrospect, this is a mixing
of responsibilities. Now, the function just checks the type of the
variable passed and returns a boolean. At the call site the value of
this boolean is used to determine whether an error message is printed.

Additionally, the function was previously autoloaded. However, because
it was called in `plugin/decfile.vim`, `autoload/decfile.vim` was
sourced when Vim started, defeating the purpose of autoloading. Now the
function has been moved to `plugin/decfile.vim` instead of being
autoloaded.